pyproject.toml
[tool.poetry]
name = "app_dash"
version = "0.1.0"
description = "Dashboard para visualização de dados de preventivas da Arpol Quality Care"
authors = ["Seu Nome <seu.email@exemplo.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
streamlit = "^1.32.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pillow = "^10.2.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
redis = "^5.0.0"
prometheus-client = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
flake8 = "^7.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
